initial cleanup for refactoring
