+++ cg/apps/demultiplex/demultiplex_api.py
@@ -3,22 +3,22 @@ import logging
 from pathlib import Path
 from typing import Dict, List, Optional
 
+from cgmodels.cg.constants import Pipeline
+from cgmodels.trailblazer.constants import AnalysisTypes
 from typing_extensions import Literal
 
-from cgmodels.trailblazer.constants import AnalysisTypes
 from cg.apps.demultiplex.sbatch import DEMULTIPLEX_COMMAND, DEMULTIPLEX_ERROR
 from cg.apps.housekeeper.hk import HousekeeperAPI
 from cg.apps.slurm.slurm_api import SlurmAPI
 from cg.apps.tb import TrailblazerAPI
 from cg.constants.constants import FileFormat
-from cg.constants.demultiplexing import DemultiplexingDirsAndFiles, BclConverter
+from cg.constants.demultiplexing import BclConverter, DemultiplexingDirsAndFiles
 from cg.constants.priority import SlurmQos
 from cg.exc import HousekeeperFileMissingError
 from cg.io.controller import WriteFile
-from cg.models.demultiplex.flow_cell import FlowCellDirectoryData
 from cg.models.demultiplex.sbatch import SbatchCommand, SbatchError
+from cg.models.flow_cell.flow_cell import FlowCellDirectoryData
 from cg.models.slurm.sbatch import Sbatch, SbatchDragen
-from cgmodels.cg.constants import Pipeline
 
 LOG = logging.getLogger(__name__)
