+++ cg/apps/demultiplex/sample_sheet/create.py
@@ -1,13 +1,14 @@
 import logging
 from typing import List
+
+from cg.apps.demultiplex.sample_sheet.models import FlowCellSample
 from cg.apps.demultiplex.sample_sheet.sample_sheet_creator import (
     SampleSheetCreator,
     SampleSheetCreatorBcl2Fastq,
     SampleSheetCreatorBCLConvert,
 )
-from cg.apps.demultiplex.sample_sheet.models import FlowCellSample
 from cg.constants.demultiplexing import BclConverter
-from cg.models.demultiplex.flow_cell import FlowCellDirectoryData
+from cg.models.flow_cell.flow_cell import FlowCellDirectoryData
 
 LOG = logging.getLogger(__name__)
