+++ tests/meta/demultiplex/test_housekeeper_storage_functions.py
@@ -15,7 +15,7 @@ from cg.meta.demultiplex.housekeeper_storage_functions import (
     add_sample_sheet_path_to_housekeeper,
 )
 from cg.models.cg_config import CGConfig
-from cg.models.demultiplex.flow_cell import FlowCellDirectoryData
+from cg.models.flow_cell.flow_cell import FlowCellDirectoryData
 
 
 def test_add_bundle_and_version_if_non_existent(demultiplex_context: CGConfig):
